requirements.txt
# requirements.txt

annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.2
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
motor==3.7.1
numpy==2.3.2
openai==1.98.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
portalocker==2.10.1
protobuf==6.31.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
PyPDF2==3.0.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
qdrant-client==1.8.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
